TileTBRec: Fix compilation with clang.
TileTBRec: Fix compilation with clang.

Gaudi property conversions.

+++ TileCalorimeter/TileTBRec/src/TileTBAANtuple.cxx
@@ -640,7 +640,7 @@ StatusCode TileTBAANtuple::storeBeamElements(const EventContext& ctx) {
 
   if ( m_completeNtuple ) {
     // Store ROD header info from collection (just from first one)
-      int nDrawersAll = m_nDrawers + m_nDrawersFlx;
+    int nDrawersAll = static_cast<unsigned int>(m_nDrawers) + static_cast<unsigned int>(m_nDrawersFlx);
     if ( collItr!=lastColl ) {
       m_l1ID.at(nDrawersAll) = (*collItr)->getLvl1Id();
       m_l1Type.at(nDrawersAll) = (*collItr)->getLvl1Type();
@@ -2450,7 +2450,7 @@ StatusCode TileTBAANtuple::initListFlx(const EventContext& ctx) {
         ATH_MSG_INFO(os.str());
 
         if (m_eventsPerFile == 0) {
-          int nDrawersAll = m_nDrawers + m_nDrawersFlx;
+          int nDrawersAll = static_cast<unsigned int>(m_nDrawers) + static_cast<unsigned int>(m_nDrawersFlx);
           m_eventsPerFile = static_cast<int>(200 / nDrawersAll) * 1000;
           ATH_MSG_INFO( "Number of events per file was 0, set it to 200k/" << nDrawersAll << " = " << m_eventsPerFile );
         }
@@ -2571,7 +2571,7 @@ void TileTBAANtuple::TRIGGER_addBranch(void)
   m_ntuplePtr->Branch("OFLunits",&m_rchUnit,"OFLunits/S");
 
   if ( m_completeNtuple ) {
-    int nDrawersAll = m_nDrawers + m_nDrawersFlx;
+    int nDrawersAll = static_cast<unsigned int>(m_nDrawers) + static_cast<unsigned int>(m_nDrawersFlx);
     if (nDrawersAll > 0) {
       m_l1ID.resize(nDrawersAll + 1);
       m_l1Type.resize(nDrawersAll + 1);
